fix: some headparts weren't managed
In 1.5.x, the internal skyrim function SkinAllGeometry internally called the
SkinSingleGeometry internal skyrim function when required.

In 1.6.x, this doesn't happen, so we need to call SkinSingleGeometry ourselves.
The algorithm proposed by igloomod (thanks :) ) processes one headpart by
headpart type, this headpart being the first found, and supposed to be the
current headpart for this type.

There was a bug in this algorithm which consisted in managing only the n first
types, n being the number of headparts. I think this resulted in some missing
headparts, in particular visible in bug reports about new races.

This bug has been first fixed by processing all the headparts.

In 1.6.353, although this fixes the invisibility bug, it made more prevalent
the missing headparts bugs, this being particularly visible in the vampire bug.

In 1.6.640, the invisibility bug has been reintroduced despite that bugfix.

This commit is a bugfix consisting in processing one headpart by headpart type,
for every headpart type. This fixes the vampire bug in my tests. The
invisibility bug is still here.