fix: fix skyrimvr build regression introduced in     7d7fa48c7669048e94b60da356f6f0bbc4924f21
The SKYRIMVR preprocessing value used in the hdtSSEUtils project, when
building the hdtSMP64 project, can't be set in hdtSMP64/Offsets.h, as
this file isn't accessible by the hdtSSEUtils project.
So the SKYRIMVR value must be set in the preprocessing tags of the
hdtSMP64 project, so that it can be used by the files in the hdtSSEUtils
project.